+++ src/lib/components/lemmy/moderation/CommentModerationMenu.svelte
@@ -20,7 +20,7 @@
     export let item: PostView | CommentView
   </script>
 
-<Menu alignment="top-center" class="top-0 h-[26px] w-[26px] ">
+<Menu alignment="top-right" class="top-0 h-[26px] w-[26px] ">
     <Button
         class="w-[26px] h-[26px] hover:!text-green-500 dark:text-zinc-400 text-slate-600"
         size="square-md"