+++ src/lib/components/ui/menu/Menu.svelte
@@ -68,7 +68,7 @@
             }}
             class="list-none absolute h-auto z-[90] overflow-auto 
                 {getOriginClass(alignment)}
-                rounded-lg py-1 min-w-[16rem] my-2 flex flex-col 
+                rounded-lg py-1 min-w-[14rem] my-2 flex flex-col 
                 shadow-md backdrop-blur-2xl border
                 bg-white  border-slate-200 
                 dark:bg-zinc-950 dark:border-zinc-800