+++ ChangeLog.md
@@ -27,6 +27,8 @@ All major/minor changes between releases will be documented here.
 - Added user option to set the number of posts shown per page
   - Note that more posts may take longer to be retrieved from your home instance.
 - Refined scroll into view when expanding/collapsing posts and post body contents. Much less "jumpy" now.
+- Comment action and moderation buttons are to the right similar to posts
+- Added debug info button to comments.
 
 ### Bugfixes
 - Youtube component now shows parameters in link